+++ internal/domain/entity/location.go
@@ -5,7 +5,7 @@ import (
 )
 
 type Location struct {
-	ID          int        `json:"id" db:"ID"`
+	ID          uint64     `json:"id" db:"ID"`
 	City        string     `json:"city" db:"City"`
 	District    string     `json:"district" db:"District"`
 	Ward        string     `json:"ward" db:"Ward"`
